fix some gcc 4.9 valid warnings
Scilab types used an int size whereas some natives functions return a
size_t. static_cast<int> is needed.

`for(iterator i = begin(), end = cases->end(); i != end; ++i)` will
disable optimizations as the end variable is not `const` thus might be
modified in the loop.

+++ scilab/modules/ast/includes/types/sparse.hxx
@@ -668,7 +668,7 @@ struct EXTERN_AST SparseBool : GenericType
 
     bool isTrue()
     {
-        if (nbTrue() == m_iSize)
+        if (static_cast<int>(nbTrue()) == m_iSize)
         {
             return true;
         }
